pyproject.toml
[project]
name = "booking-system-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "requests>=2.31.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.3",
    "ruff>=0.12.7",
    "ty>=0.0.1a16",
]
